Revert "refactor: move ``GetRelease`` to setup package"
This reverts commit 7f4f48bf902deaf60409186b45545e48eebff7c5.
